Substantial changes to codebase
- Split out API layer separately
- Entity now inherits interface from API class
- Schema now split out separately, also inherits interface
- Entity now not dependent on Client
- Split out dataclasses, errors, api, schema and entity in separate files
