+++ dataverse_api/errors.py
@@ -1,0 +1,15 @@
+"""
+Collection of Error classes specific to Dataverse API.
+"""
+
+
+class DataverseError(Exception):
+    def __init__(self, message: str, status_code=None, response=None) -> None:
+        super().__init__(message)
+        self.status_code = status_code
+        self.response = response
+
+
+class DataverseValidationError(DataverseError):
+    def __init__(self, message: str, status_code=None, response=None) -> None:
+        super().__init__(message, status_code, response)
